requirements.txt
Flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
requests==2.31.0
python-dotenv==1.0.0  # Optional, for .env file support